requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.99.1
filelock==3.18.0
frozenlist==1.7.0
h11==0.16.0
idna==3.10
Jinja2==3.1.2
joblib==1.4.2
MarkupSafe==3.0.2
motor==3.7.0
multidict==6.6.4
numpy==1.24.3
pandas==2.0.2
passlib==1.7.4
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.9
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
python-whois==0.9.5
pytz==2025.2
requests==2.31.0
requests-file==2.1.0
rsa==4.9.1
scikit-learn==1.2.2
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.27.0
threadpoolctl==3.6.0
tldextract==5.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.22.0
xgboost==1.7.6
yarl==1.20.1
